+++ src/__playground__/index.ts
@@ -1,1 +1,9 @@
-console.log('hello world');
+import { define } from '../index.js';
+
+const api = await define({
+    add: () => {
+        return 1 + 1;
+    },
+});
+
+await api({ name: 'add' });