+++ src/__playground__/worker.ts
@@ -1,0 +1,7 @@
+import { define } from '../index.js';
+
+export const api = await define({
+    add: () => {
+        return 1 + 1;
+    },
+});